requirements.txt
Django==3.1.13
pytz==2019.3
sqlparse==0.3.1
gunicorn


python-dotenv==0.21.0
mysqlclient==2.1.1
django-ckeditor==5.9.0